Updated for Server 2016
Quick change as reported privately
----
Quote:
Today I ran into a problem when executing your script on Windows Server 2016, it fails with error "servermanagercmd is not recognized as internal or external command..."
The problem is that it incorrectly identifies Server 2016 as Server 2008.
The Major version number of Server 2016 is 10, which returns true for the first clause. The script then goes into checking the minor version, and since the minor version number is 0, it returns false for all clauses, until it runs into the Server 2008 elseif and tries to run the unsopported command.